requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
python-linkedin==4.1.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
langchain==0.0.350
python-twitter==3.5.0
PyGithub==2.1.1 